style: spotless apply refactor: remove convention plugins.
