+++ compose/build.gradle.kts
@@ -1,5 +1,6 @@
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTarget
 import org.jetbrains.kotlin.gradle.targets.js.dsl.ExperimentalWasmDsl
+import java.util.Properties
 
 // Copyright 2023, Sergei Gagarin and the project contributors
 // SPDX-License-Identifier: Apache-2.0
@@ -8,7 +9,8 @@ plugins {
     alias(libs.plugins.multiplatform)
     alias(libs.plugins.compose)
     alias(libs.plugins.android.library)
-    id("convention.publication")
+    id("maven-publish")
+    id("signing")
 }
 
 group = "tech.annexflow.compose"
@@ -121,4 +123,85 @@ android {
         sourceCompatibility = JavaVersion.toVersion(rootProject.extra.get("jvmTarget") as String)
         targetCompatibility = JavaVersion.toVersion(rootProject.extra.get("jvmTarget") as String)
     }
+}
+
+ext["signing.keyId"] = null
+ext["signing.password"] = null
+ext["signing.secretKeyRingFile"] = null
+ext["ossrhUsername"] = null
+ext["ossrhPassword"] = null
+
+val secretPropsFile = project.rootProject.file("local.properties")
+if (secretPropsFile.exists()) {
+    secretPropsFile.reader().use {
+        Properties().apply { load(it) }
+    }.onEach { (name, value) ->
+        ext[name.toString()] = value
+    }
+} else {
+    ext["signing.keyId"] = System.getenv("SIGNING_KEY_ID")
+    ext["signing.password"] = System.getenv("SIGNING_PASSWORD")
+    ext["signing.secretKeyRingFile"] = System.getenv("SIGNING_SECRET_KEY_RING_FILE")
+    ext["ossrhUsername"] = System.getenv("OSSRH_USERNAME")
+    ext["ossrhPassword"] = System.getenv("OSSRH_PASSWORD")
+}
+
+val javadocJar by tasks.registering(Jar::class) {
+    archiveClassifier.set("javadoc")
+}
+
+fun getExtraString(name: String) = ext[name]?.toString()
+
+publishing {
+    repositories {
+        maven {
+            name = "sonatype"
+            setUrl("https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/")
+            credentials {
+                username = getExtraString("ossrhUsername")
+                password = getExtraString("ossrhPassword")
+            }
+        }
+    }
+
+    publications.withType<MavenPublication> {
+        artifact(javadocJar.get())
+
+        pom {
+            artifactId = "constraintlayout-compose-multiplatform"
+            name.set("Compose ConstraintLayout")
+            description.set("Provides Compose ConstraintLayout for building responsive UIs")
+            url.set("https://github.com/lavmee/constraintlayout-compose-multiplatform/")
+            inceptionYear.set("2023")
+
+            licenses {
+                license {
+                    name.set("The Apache Software License, Version 2.0")
+                    url.set("=http://www.apache.org/licenses/LICENSE-2.0.txt")
+                    distribution.set("repo")
+                }
+            }
+            developers {
+                developer {
+                    id.set("lavmee")
+                    name.set("Sergei Gagarin")
+                    email.set("lavmeejetski@gmail.com")
+                }
+            }
+            scm {
+                url.set("scm:git:git://github.com/lavmee/constraintlayout-compose-multiplatform.git")
+            }
+        }
+    }
+}
+
+signing {
+    if (getExtraString("signing.keyId") != null) {
+        sign(publishing.publications)
+    }
+}
+
+val signingTasks = tasks.withType<Sign>()
+tasks.withType<AbstractPublishToMaven>().configureEach {
+    mustRunAfter(signingTasks)
 }
