--- convention-plugins/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    `kotlin-dsl` // Is needed to turn our build logic written in Kotlin into Gralde Plugin
-}
-
-repositories {
-    gradlePluginPortal() // To use 'maven-publish' and 'signing' plugins in our own plugin
-}
