--- convention-plugins/src/main/kotlin/convention.publication.gradle.kts
@@ -1,87 +1,0 @@
-import java.util.Properties
-
-plugins {
-    id("maven-publish")
-    id("signing")
-}
-
-ext["signing.keyId"] = null
-ext["signing.password"] = null
-ext["signing.secretKeyRingFile"] = null
-ext["ossrhUsername"] = null
-ext["ossrhPassword"] = null
-
-val secretPropsFile = project.rootProject.file("local.properties")
-if (secretPropsFile.exists()) {
-    secretPropsFile.reader().use {
-        Properties().apply { load(it) }
-    }.onEach { (name, value) ->
-        ext[name.toString()] = value
-    }
-} else {
-    ext["signing.keyId"] = System.getenv("SIGNING_KEY_ID")
-    ext["signing.password"] = System.getenv("SIGNING_PASSWORD")
-    ext["signing.secretKeyRingFile"] = System.getenv("SIGNING_SECRET_KEY_RING_FILE")
-    ext["ossrhUsername"] = System.getenv("OSSRH_USERNAME")
-    ext["ossrhPassword"] = System.getenv("OSSRH_PASSWORD")
-}
-
-val javadocJar by tasks.registering(Jar::class) {
-    archiveClassifier.set("javadoc")
-}
-
-fun getExtraString(name: String) = ext[name]?.toString()
-
-publishing {
-    repositories {
-        maven {
-            name = "sonatype"
-            setUrl("https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/")
-            credentials {
-                username = getExtraString("ossrhUsername")
-                password = getExtraString("ossrhPassword")
-            }
-        }
-    }
-
-    publications.withType<MavenPublication> {
-        artifact(javadocJar.get())
-
-        pom {
-            artifactId = "constraintlayout-compose-multiplatform"
-            name.set("Compose ConstraintLayout")
-            description.set("Provides Compose ConstraintLayout for building responsive UIs")
-            url.set("https://github.com/lavmee/constraintlayout-compose-multiplatform/")
-            inceptionYear.set("2023")
-
-            licenses {
-                license {
-                    name.set("The Apache Software License, Version 2.0")
-                    url.set("=http://www.apache.org/licenses/LICENSE-2.0.txt")
-                    distribution.set("repo")
-                }
-            }
-            developers {
-                developer {
-                    id.set("lavmee")
-                    name.set("Sergei Gagarin")
-                    email.set("lavmeejetski@gmail.com")
-                }
-            }
-            scm {
-                url.set("scm:git:git://github.com/lavmee/constraintlayout-compose-multiplatform.git")
-            }
-        }
-    }
-}
-
-signing {
-    if (getExtraString("signing.keyId") != null) {
-        sign(publishing.publications)
-    }
-}
-
-val signingTasks = tasks.withType<Sign>()
-tasks.withType<AbstractPublishToMaven>().configureEach {
-    mustRunAfter(signingTasks)
-}
