+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ compose = "1.6.0-rc02"
 compose-uitooling = "1.6.1"
 kotlin = "1.9.22"
 agp = "8.2.2"
-spotless = "6.23.3"
+spotless = "6.25.0"
 
 [libraries]
